requirements.txt
anthropic>=0.25.0
supabase>=2.0.0
sendgrid>=6.11.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytest>=8.0.0
pytest-mock>=3.12.0
